+++ changelog.md
@@ -2,7 +2,10 @@
 
 
 ## `v2.0.0`
-- **breaking change** moved github.com/Azure/azure-amqp-common-go/persist to github.com/Azure/azure-event-hubs-go/persist
+- **breaking change:** moved github.com/Azure/azure-amqp-common-go/persist to
+  github.com/Azure/azure-event-hubs-go/persist
+- **breaking change:** changed batch message sending to use a safe batch iterator rather than leaving batch sizing to
+  the consumer.
 - move tracing to devigned/tab so to not have to take a direct dependency on opentracing or opencensus
 
 ## `v1.3.1`